Added a program that grabs everything!
I made a program called makeAPeel.py
(I though that was the name of our project lol my fault)
When you run the file it will ask for a link. Just grab the link at the
top of the youtube video page and paste it in (Make sure the video has
captions). It will then make turn the youtube transcript into a
temporary transcript text file and then it will run both of Irene's and
Casey's code. After it finishes it will then delete the temporary file.